requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
anthropic==0.71.0
python-multipart==0.0.6
aiofiles==23.2.1
olefile==0.47
jinja2==3.1.6
bcrypt==4.1.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]>=2.5.0
pydantic-core>=2.14.1
email-validator>=2.0.0
